Some todo app methods converted to async await

+++ node-todo-api/server/server.js
@@ -60,25 +60,24 @@ app.post('/todos', authenticate, (req, res) => {
     });
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id;
+app.delete('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id;
 
     if (!ObjectID.isValid(id)) {
         return res.status(404).send();
     }
+    
+    try {
+        const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id })
 
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
         if (todo === null) {
             return res.status(404).send();
         }
 
-        res.send({todo});
-    }, (err) => {
-        res.status(400).send(err);
-    });
+        res.send({ todo });
+    } catch (error) {
+        res.status(400).send(error.message);
+    };
 });
 
 app.patch('/todos/:id', authenticate, (req, res) => {
@@ -110,43 +109,43 @@ app.patch('/todos/:id', authenticate, (req, res) => {
     });
 });
 
-app.post('/users', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password']);
-    let user = new User(body);
+app.post('/users', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = new User(body);
 
-    user.save().then(() => {
-        return user.generateAuthToken();
-    }).then((token) => {
+        await user.save();
+        const token = await user.generateAuthToken();
+
         // 'x-' prefix means the custom header.
         res.header('x-auth', token).send(user);
-    })
-    .catch((err) => {
-        res.status(400).send(err);
-    });
+    } catch (error) {
+        res.status(400).send(error);
+    };
 });
 
 app.get('/users/me', authenticate, (req, res) => {
     res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password']);
-
-    User.findByCredentials(body.email, body.password).then((user) => {
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        });
-    }).catch((err) => {
+app.post('/users/login', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (error) {
         res.status(400).send();
-    });
+    };
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        const result = await req.user.removeToken(req.token);
         res.status(200).send();
-    }).catch((err) => {
+    } catch (error) {
         res.status(400).send();
-    });
+    };
 });
 
 app.listen(port, () => {
